+++ relay-server/src/actors/global_config.rs
@@ -8,46 +8,58 @@ use crate::actors::processor::EnvelopeProcessor;
 use crate::actors::project_upstream::GetProjectStates;
 use crate::actors::upstream::{SendQuery, UpstreamRelay};
 
-/// Service implementing the [`GlobalConfigResponse`] interface.
+/// Service implementing the [`GlobalConfiguration`] interface.
 ///
 /// The service is responsible for fetching the global config and
-/// forwarding it to the services that require it.
+/// forwarding it to the services that require it, and for serving downstream relays.
 #[derive(Debug)]
-pub struct GlobalConfigService {
+pub struct GlobalConfigurationService {
     global_config: Arc<GlobalConfig>,
-    project_cache: Addr<EnvelopeProcessor>,
+    envelope_processor: Addr<EnvelopeProcessor>,
     upstream: Addr<UpstreamRelay>,
 }
 
 /// Service interface for the [`GetGlobalConfig`] message.
-pub struct GlobalConfigResponse(Sender<Arc<GlobalConfig>>);
+pub enum GlobalConfiguration {
+    GetGlobalConfig(Sender<Arc<GlobalConfig>>),
+}
 
+impl Interface for GlobalConfiguration {}
+
+/// The message for requesting the most recent global config from [`GlobalConfigurationService`]
 pub struct GetGlobalConfig;
 
-impl FromMessage<GetGlobalConfig> for GlobalConfigResponse {
+impl FromMessage<GetGlobalConfig> for GlobalConfiguration {
     type Response = AsyncResponse<Arc<GlobalConfig>>;
 
-    fn from_message(_message: GetGlobalConfig, sender: Sender<Arc<GlobalConfig>>) -> Self {
-        Self(sender)
+    fn from_message(_: GetGlobalConfig, sender: Sender<Arc<GlobalConfig>>) -> Self {
+        Self::GetGlobalConfig(sender)
     }
 }
 
-impl Interface for GlobalConfigResponse {}
-
-impl GlobalConfigService {
-    pub fn new(project_cache: Addr<EnvelopeProcessor>, upstream: Addr<UpstreamRelay>) -> Self {
+impl GlobalConfigurationService {
+    /// Creates a new [`GlobalConfigurationService`]
+    pub fn new(envelope_processor: Addr<EnvelopeProcessor>, upstream: Addr<UpstreamRelay>) -> Self {
         let global_config = Arc::new(GlobalConfig::default());
         Self {
             global_config,
-            project_cache,
+            envelope_processor,
             upstream,
+        }
+    }
+
+    fn handle_message(&self, message: GlobalConfiguration) {
+        match message {
+            GlobalConfiguration::GetGlobalConfig(sender) => {
+                sender.send(Arc::clone(&self.global_config));
+            }
         }
     }
 
     /// Forwards the given global config to the services that require it.
     async fn update_global_config(&mut self) {
         let upstream_relay: Addr<UpstreamRelay> = self.upstream.clone();
-        let project_cache = self.project_cache.clone();
+        let envelope_processor = self.envelope_processor.clone();
 
         let query = GetProjectStates {
             public_keys: vec![],
@@ -57,13 +69,13 @@ impl GlobalConfigService {
         };
 
         match upstream_relay.send(SendQuery(query)).await {
-            Ok(Ok(response)) => match response.global {
+            Ok(Ok(response)) => match response.global_config {
                 Some(global_config) => {
                     let global_config = Arc::new(global_config);
                     self.global_config = global_config.clone();
-                    project_cache.send::<Arc<GlobalConfig>>(global_config);
+                    envelope_processor.send::<Arc<GlobalConfig>>(global_config);
                 }
-                None => relay_log::error!("Global config request returned a None value"),
+                None => relay_log::error!("Upstream response didn't include a global config"),
             },
             Err(e) => {
                 relay_log::error!("failed to send global config request: {}", e);
@@ -75,27 +87,23 @@ impl GlobalConfigService {
     }
 }
 
-impl Service for GlobalConfigService {
-    type Interface = GlobalConfigResponse;
+impl Service for GlobalConfigurationService {
+    type Interface = GlobalConfiguration;
 
     fn spawn_handler(mut self, mut rx: relay_system::Receiver<Self::Interface>) {
         tokio::spawn(async move {
             let mut ticker = tokio::time::interval(Duration::from_secs(10));
-            relay_log::info!("global config started");
+            relay_log::info!("global configuration service started");
 
             loop {
                 tokio::select! {
                     biased;
                     _ = ticker.tick() => self.update_global_config().await,
-                    message = rx.recv() => {
-                        if let Some(sender) = message {
-                            sender.0.send(self.global_config.clone());
-                        }
-                    },
+                    Some(message) = rx.recv() => self.handle_message(message),
                     else => break,
                 }
             }
-            relay_log::info!("global config stopped");
+            relay_log::info!("global configuration service stopped");
         });
     }
 }